+++ src/app/views/common/lazy-loader/component-registry/index.tsx
@@ -10,7 +10,7 @@ import LazyCopyButton from  '../../copy-button/CopyButtonV9';
 import LazyAuth from  '../../../query-runner/request/auth/AuthV9';
 import LazyRequestHeaders from  '../../../query-runner/request/headers/RequestHeadersV9';
 import LazyHistory from  '../../../sidebar/history/History';
-import LazyResourceExplorer from  '../../../sidebar/resource-explorer/ResourceExplorer';
+import LazyResourceExplorer from  '../../../sidebar/resource-explorer/ResourceExplorerV9';
 
 export const Permissions =() => {
   return (
